+++ src/writer/basic.rs
@@ -129,7 +129,7 @@ impl Basic {
     /// Outputs [started] [`Feature`] to STDOUT.
     ///
     /// [started]: event::Feature::Started
-    /// [`Feature`]: [`gherkin::Feature`]
+    /// [`Feature`]: gherkin::Feature
     fn feature_started(&mut self, feature: &gherkin::Feature) {
         self.lines_to_clear = 1;
         self.write_line(
@@ -169,7 +169,7 @@ impl Basic {
     /// Outputs [started] [`Rule`] to STDOUT.
     ///
     /// [started]: event::Rule::Started
-    /// [`Rule`]: [`gherkin::Rule`]
+    /// [`Rule`]: gherkin::Rule
     fn rule_started(&mut self, rule: &gherkin::Rule) {
         self.lines_to_clear = 1;
         self.indent += 2;
@@ -187,17 +187,28 @@ impl Basic {
     /// [background]: event::Background
     /// [started]: event::Scenario::Started
     /// [step]: event::Step
+    /// [`Scenario`]: gherkin::Scenario
     fn scenario<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         scenario: &gherkin::Scenario,
         ev: &event::Scenario<W>,
     ) {
-        use event::Scenario;
+        use event::{Hook, Scenario};
 
         match ev {
             Scenario::Started => {
                 self.scenario_started(scenario);
+            }
+            Scenario::Hook(_, Hook::Started) => {
+                self.indent += 4;
+            }
+            Scenario::Hook(which, Hook::Failed(world, info)) => {
+                self.hook_failed(feat, scenario, *which, world.as_ref(), info);
+                self.indent = self.indent.saturating_sub(4);
+            }
+            Scenario::Hook(_, Hook::Passed) => {
+                self.indent = self.indent.saturating_sub(4);
             }
             Scenario::Background(bg, ev) => {
                 self.background(feat, bg, ev);
@@ -209,10 +220,41 @@ impl Basic {
         }
     }
 
+    /// Outputs [failed] [`Scenario`]'s hook to STDOUT.
+    ///
+    /// [failed]: event::Hook::Failed
+    /// [`Scenario`]: gherkin::Scenario
+    fn hook_failed<W: Debug>(
+        &mut self,
+        feat: &gherkin::Feature,
+        sc: &gherkin::Scenario,
+        which: event::HookType,
+        world: Option<&W>,
+        info: &Info,
+    ) {
+        self.clear_last_lines_if_term_present();
+
+        self.write_line(&self.styles.err(format!(
+            "{indent}\u{2718}  Scenario's {} hook failed {}:{}:{}\n\
+             {indent}   Captured output: {}{}",
+            which,
+            feat.path
+                .as_ref()
+                .and_then(|p| p.to_str())
+                .unwrap_or(&feat.name),
+            sc.position.line,
+            sc.position.col,
+            coerce_error(info),
+            format_world(world, self.indent.saturating_sub(3) + 3),
+            indent = " ".repeat(self.indent.saturating_sub(3)),
+        )))
+        .unwrap();
+    }
+
     /// Outputs [started] [`Scenario`] to STDOUT.
     ///
     /// [started]: event::Scenario::Started
-    /// [`Scenario`]: [`gherkin::Scenario`]
+    /// [`Scenario`]: gherkin::Scenario
     fn scenario_started(&mut self, scenario: &gherkin::Scenario) {
         self.lines_to_clear = 1;
         self.indent += 2;
@@ -231,7 +273,7 @@ impl Basic {
     /// [passed]: event::Step::Passed
     /// [skipped]: event::Step::Skipped
     /// [started]: event::Step::Started
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Step`]: gherkin::Step
     fn step<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
@@ -268,7 +310,7 @@ impl Basic {
     /// [passed]: event::Step::Passed
     /// [skipped]: event::Step::Skipped
     /// [started]: event::Step::Started
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Step`]: gherkin::Step
     fn step_started(&mut self, step: &gherkin::Step) {
         self.indent += 4;
         if self.styles.is_present {
@@ -290,7 +332,7 @@ impl Basic {
     /// Outputs [passed] [`Step`] to STDOUT.
     ///
     /// [passed]: event::Step::Passed
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Step`]: gherkin::Step
     fn step_passed(
         &mut self,
         step: &gherkin::Step,
@@ -325,7 +367,7 @@ impl Basic {
     /// Outputs [skipped] [`Step`] to STDOUT.
     ///
     /// [skipped]: event::Step::Skipped
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Step`]: gherkin::Step
     fn step_skipped(&mut self, feat: &gherkin::Feature, step: &gherkin::Step) {
         self.clear_last_lines_if_term_present();
         self.write_line(&self.styles.skipped(format!(
@@ -351,7 +393,7 @@ impl Basic {
     /// Outputs [failed] [`Step`] to STDOUT.
     ///
     /// [failed]: event::Step::Failed
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Step`]: gherkin::Step
     fn step_failed<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
@@ -413,8 +455,8 @@ impl Basic {
     /// [passed]: event::Step::Passed
     /// [skipped]: event::Step::Skipped
     /// [started]: event::Step::Started
-    /// [`Background`]: [`gherkin::Background`]
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Background`]: gherkin::Background
+    /// [`Step`]: gherkin::Step
     fn background<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
@@ -451,8 +493,8 @@ impl Basic {
     /// [passed]: event::Step::Passed
     /// [skipped]: event::Step::Skipped
     /// [started]: event::Step::Started
-    /// [`Background`]: [`gherkin::Background`]
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Background`]: gherkin::Background
+    /// [`Step`]: gherkin::Step
     fn bg_step_started(&mut self, step: &gherkin::Step) {
         self.indent += 4;
         if self.styles.is_present {
@@ -474,8 +516,8 @@ impl Basic {
     /// Outputs [passed] [`Background`] [`Step`] to STDOUT.
     ///
     /// [passed]: event::Step::Passed
-    /// [`Background`]: [`gherkin::Background`]
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Background`]: gherkin::Background
+    /// [`Step`]: gherkin::Step
     fn bg_step_passed(
         &mut self,
         step: &gherkin::Step,
@@ -510,8 +552,8 @@ impl Basic {
     /// Outputs [skipped] [`Background`] [`Step`] to STDOUT.
     ///
     /// [skipped]: event::Step::Skipped
-    /// [`Background`]: [`gherkin::Background`]
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Background`]: gherkin::Background
+    /// [`Step`]: gherkin::Step
     fn bg_step_skipped(
         &mut self,
         feat: &gherkin::Feature,
@@ -541,8 +583,8 @@ impl Basic {
     /// Outputs [failed] [`Background`] [`Step`] to STDOUT.
     ///
     /// [failed]: event::Step::Failed
-    /// [`Background`]: [`gherkin::Background`]
-    /// [`Step`]: [`gherkin::Step`]
+    /// [`Background`]: gherkin::Background
+    /// [`Step`]: gherkin::Step
     fn bg_step_failed<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,