+++ src/writer/mod.rs
@@ -83,6 +83,12 @@ pub trait Failure<World>: Writer<World> {
     /// Returns number of parsing errors.
     #[must_use]
     fn parsing_errors(&self) -> usize;
+
+    /// Returns number of failed [`Scenario`] hooks.
+    ///
+    /// [`Scenario`]: gherkin::Scenario
+    #[must_use]
+    fn hook_errors(&self) -> usize;
 }
 
 /// Extension of [`Writer`] allowing its normalization and summarization.