+++ src/writer/normalized.rs
@@ -134,6 +134,10 @@ where
     fn parsing_errors(&self) -> usize {
         self.writer.parsing_errors()
     }
+
+    fn hook_errors(&self) -> usize {
+        self.writer.hook_errors()
+    }
 }
 
 /// Normalization queue for incoming events.